new env params and fix for dbtools
added settings for not limitting AH  and not expiring auctions. This needs to be fixed later as this makes it friendly for our server but not for other people that want to test.

dbtools.py was having issues as the root of the docker settings wasnt being copied as it needs to get the hash of the current version. now we are copying the root instead of just the server folder.